# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Three Kingdoms City Node",
    "version": "1.3.0"
  },
  "paths": {
...
    "/action/work": {
      "post": {
        "tags": [
          "action"
        ],
        "summary": "Action Work",
        "operationId": "action_work_action_work_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/action/train": {
      "post": {
        "tags": [
          "action"
        ],
        "summary": "Action Train",
        "operationId": "action_train_action_train_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/action/attack": {
      "post": {
        "tags": [
          "action"
        ],
        "summary": "Action Attack",
        "operationId": "action_attack_action_attack_post",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Id"
            }
          },
          {
            "name": "target_agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Target Agent Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrainActionRequest": {
        "properties": {
          "agent_id": {
            "type": "integer",
            "title": "Agent Id"
          },
          "troop_type": {
            "type": "string",
            "title": "Troop Type"
          },
          "quantity": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "troop_type",
          "quantity"
        ],
        "title": "TrainActionRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkActionRequest": {
        "properties": {
          "agent_id": {
            "type": "integer",
            "title": "Agent Id"
          },
          "task": {
            "type": "string",
            "title": "Task"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "task"
        ],
        "title": "WorkActionRequest"
      }
    }
  }
}